[projectfirma/#1238] Control display order of Taxonomy and Classification Items Relocate EditSortOrder* files to the shared views namespace/folder
