+++ Source/ProjectFirma.Web/Global.asax.cs
@@ -56,7 +56,8 @@ namespace ProjectFirma.Web
                 "~/Views/Shared/ProjectLocationControls/{0}.cshtml",
                 "~/Views/Shared/ProjectWatershedControls/{0}.cshtml",
                 "~/Views/Shared/ProjectUpdateDiffControls/{0}.cshtml",
-                "~/Views/Shared/ProjectOrganization/{0}.cshtml"
+                "~/Views/Shared/ProjectOrganization/{0}.cshtml",
+                "~/Views/Shared/SortOrder/{0}.cshtml"
             } };
             // read the log4net configuration from the web.config file
             XmlConfigurator.Configure();
