+++ Source/ProjectFirma.Web/ProjectFirma.Web.csproj
@@ -46,9 +46,9 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Views\Classification\EditSortOrderViewModel.cs" />
-    <Compile Include="Views\Classification\EditSortOrder.cs" />
-    <Compile Include="Views\Classification\EditSortOrderViewData.cs" />
+    <Compile Include="Views\Shared\SortOrder\EditSortOrderViewModel.cs" />
+    <Compile Include="Views\Shared\SortOrder\EditSortOrder.cs" />
+    <Compile Include="Views\Shared\SortOrder\EditSortOrderViewData.cs" />
     <Compile Include="Models\Generated\AssessmentGoal.Binding.cs" />
     <Compile Include="Models\Generated\AssessmentGoal.DatabaseContextExtensions.cs" />
     <Compile Include="Models\Generated\AssessmentGoalPrimaryKey.Binding.cs" />
@@ -734,7 +734,7 @@
     <Content Include="Views\CustomPage\CustomPageDetails.cshtml" />
     <Content Include="Views\CustomPage\Index.cshtml" />
     <Content Include="Views\CustomPage\Edit.cshtml" />
-    <Content Include="Views\Classification\EditSortOrder.cshtml" />
+    <Content Include="Views\Shared\SortOrder\EditSortOrder.cshtml" />
     <None Include="Views\PerformanceMeasure\EditSubcategoriesAndOptions.cshtml" />
     <Compile Include="Models\NotificationType.cs" />
     <Compile Include="Models\OrganizationType.cs" />
