+++ Source/ProjectFirma.Web/Views/Shared/SortOrder/EditSortOrder.cshtml
@@ -20,7 +20,7 @@ Source code is available upon request via <support@sitkatech.com>.
 -----------------------------------------------------------------------*@
 @using LtInfo.Common.HtmlHelperExtensions
 @using LtInfo.Common.Mvc
-@inherits ProjectFirma.Web.Views.Classification.EditSortOrder
+@inherits ProjectFirma.Web.Views.Shared.SortOrder.EditSortOrder
 
 <script src="~/Content/Sortable.js"></script>
